Large Util Refactor
