+++ src/main/java/dev/toastmc/client/command/Hide.kt
@@ -4,13 +4,12 @@ import com.mojang.brigadier.CommandDispatcher
 import dev.toastmc.client.command.util.*
 import dev.toastmc.client.command.util.type.ModuleArgumentType
 import dev.toastmc.client.module.Module
-import dev.toastmc.client.util.MessageUtil
+import dev.toastmc.client.util.Color
+import dev.toastmc.client.util.sendMessage
 import net.minecraft.server.command.CommandSource
 import net.minecraft.util.Formatting
 
-@Cmd(
-        name = "hide"
-)
+@Cmd(name = "hide")
 class Hide : Command() {
     override fun register(dispatcher: CommandDispatcher<CommandSource>) {
         dispatcher register rootLiteral("hide") {
@@ -18,7 +17,7 @@ class Hide : Command() {
                 does { ctx ->
                     val mod: Module = "module" from ctx
                     mod.hidden = !mod.hidden
-                    MessageUtil.sendMessage("${mod.label} is now being ${if (mod.hidden) Formatting.RED.toString() + "HIDDEN" else Formatting.GREEN.toString() + "SHOWN"}", MessageUtil.Color.GRAY)
+                    sendMessage("${mod.label} is now being ${if (mod.hidden) Formatting.RED.toString() + "HIDDEN" else Formatting.GREEN.toString() + "SHOWN"}", Color.GRAY)
                     0
                 }
             }
