+++ src/main/java/dev/toastmc/client/command/Pos.kt
@@ -2,22 +2,21 @@ package dev.toastmc.client.command
 
 import com.mojang.brigadier.CommandDispatcher
 import dev.toastmc.client.command.util.*
-import dev.toastmc.client.util.MessageUtil
+import dev.toastmc.client.util.Color
 import dev.toastmc.client.util.mc
+import dev.toastmc.client.util.sendMessage
 import net.minecraft.client.network.ClientPlayerEntity
 import net.minecraft.server.command.CommandSource
 import net.minecraft.util.Formatting
 import java.text.DecimalFormat
 
-@Cmd(
-        name = "pos"
-)
+@Cmd(name = "pos")
 class Pos : Command() {
     override fun register(dispatcher: CommandDispatcher<CommandSource>) {
         dispatcher register rootLiteral("pos") {
             does {
                 mc.keyboard.clipboard = formatPlayerCoords(mc.player)
-                MessageUtil.sendMessage("${Formatting.RED}!!${Formatting.GREEN} Coordinates saved to clipboard ${Formatting.RED}!!", MessageUtil.Color.GREEN)
+                sendMessage("${Formatting.RED}!!${Formatting.GREEN} Coordinates saved to clipboard ${Formatting.RED}!!", Color.GREEN)
                 0
             }
         }
