+++ src/main/java/dev/toastmc/client/command/Prefix.kt
@@ -3,7 +3,8 @@ package dev.toastmc.client.command
 import com.mojang.brigadier.CommandDispatcher
 import dev.toastmc.client.ToastClient
 import dev.toastmc.client.command.util.*
-import dev.toastmc.client.util.MessageUtil
+import dev.toastmc.client.util.Color
+import dev.toastmc.client.util.sendMessage
 import net.minecraft.server.command.CommandSource
 
 @Cmd(
@@ -17,7 +18,7 @@ class Prefix : Command() {
                     val str: String = "char" from ctx
                     val char: Char = str.toCharArray()[0]
                     ToastClient.CMD_PREFIX = char.toString()
-                    MessageUtil.sendMessage("Prefix has been set to: $char", MessageUtil.Color.GREEN)
+                    sendMessage("Prefix has been set to: $char", Color.GREEN)
                     0
                 }
             }
