+++ src/main/java/dev/toastmc/client/command/Toggle.kt
@@ -4,7 +4,8 @@ import com.mojang.brigadier.CommandDispatcher
 import dev.toastmc.client.command.util.*
 import dev.toastmc.client.command.util.type.ModuleArgumentType
 import dev.toastmc.client.module.Module
-import dev.toastmc.client.util.MessageUtil
+import dev.toastmc.client.util.Color
+import dev.toastmc.client.util.sendMessage
 import net.minecraft.server.command.CommandSource
 import net.minecraft.util.Formatting
 
@@ -18,7 +19,7 @@ class Toggle : Command() {
                 does { ctx ->
                     val mod: Module = "module" from ctx
                     mod.enabled = !mod.enabled
-                    MessageUtil.sendMessage("Toggled ${mod.label}${if (!mod.enabled) Formatting.RED.toString() + " OFF" else Formatting.GREEN.toString() + " ON"}", MessageUtil.Color.GRAY)
+                    sendMessage("Toggled ${mod.label}${if (!mod.enabled) Formatting.RED.toString() + " OFF" else Formatting.GREEN.toString() + " ON"}", Color.GRAY)
                     0
                 }
             }
