+++ src/main/java/dev/toastmc/client/module/movement/AutoWalk.kt
@@ -1,12 +1,10 @@
 package dev.toastmc.client.module.movement
 
-import baritone.api.pathing.goals.GoalBlock
 import dev.toastmc.client.ToastClient.Companion.EVENT_BUS
 import dev.toastmc.client.event.PacketEvent
 import dev.toastmc.client.module.Category
 import dev.toastmc.client.module.Module
 import dev.toastmc.client.module.ModuleManifest
-import dev.toastmc.client.util.Baritone
 import dev.toastmc.client.util.mc
 import me.zero.alpine.listener.EventHandler
 import me.zero.alpine.listener.EventHook
@@ -29,33 +27,14 @@ class AutoWalk : Module() {
     override fun onDisable() {
         super.onDisable()
         EVENT_BUS.unsubscribe(onEvent)
-        Baritone.invoke {
-            if(Baritone.baritonePresent()){
-                if (Baritone.BARITONE.customGoalProcess.isActive){
-                    Baritone.BARITONE.customGoalProcess.setGoalAndPath(null)
-                }
-            }
-        }
     }
 
     @EventHandler
     private val onEvent = Listener(EventHook<PacketEvent.Send> {
         if (mc.player == null) return@EventHook
-        var baritone = false
-        Baritone.invoke {
-            if(Baritone.baritonePresent()){
-                if (!Baritone.BARITONE.customGoalProcess.isActive){
-                    Baritone.BARITONE.customGoalProcess.setGoalAndPath(GoalBlock(mc.player!!.blockPos.method_30513(mc.player!!.movementDirection.axis, 1000)))
-                }
-            }
-            baritone = true
-            return@invoke
-        }
-        if (!baritone) {
-            if (it.packet is PlayerMoveC2SPacket || it.packet is InventoryS2CPacket) {
-                mc.options.sprintToggled = true
-                mc.options.keyForward.isPressed = true
-            }
+        if (it.packet is PlayerMoveC2SPacket || it.packet is InventoryS2CPacket) {
+            mc.options.sprintToggled = true
+            mc.options.keyForward.isPressed = true
         }
     })
 }
