--- src/main/java/dev/toastmc/client/util/ASCII.kt
@@ -1,29 +1,0 @@
-package dev.toastmc.client.util
-
-object ASCII {
-    /*
-    Added by RemainingToast - 1/06/20
-     */
-    @JvmStatic
-    fun printFancyConsoleMSG() {
-        println("_________________________________________________________")
-        println("     _________ _______  _______  _______ _________           ")
-        println("     \\__   __/(  ___  )(  ___  )(  ____ \\\\__   __/        ")
-        println("        ) (   | (   ) || (   ) || (    \\/   ) (             ")
-        println("        | |   | |   | || (___) || (_____    | |              ")
-        println("        | |   | |   | ||  ___  |(_____  )   | |              ")
-        println("        | |   | |   | || (   ) |      ) |   | |              ")
-        println("        | |   | (___) || )   ( |/\\____) |   | |             ")
-        println("        )_(   (_______)|/     \\|\\_______)   )_(            ")
-        println("  _______  _       _________ _______  _       _________  ")
-        println(" (  ____ \\( \\      \\__   __/(  ____ \\( (    /|\\__   __/")
-        println(" | (    \\/| (         ) (   | (    \\/|  \\  ( |   ) (   ")
-        println(" | |      | |         | |   | (__    |   \\ | |   | |   ")
-        println(" | |      | |         | |   |  __)   | (\\ \\) |   | |   ")
-        println(" | |      | |         | |   | (      | | \\   |   | |   ")
-        println(" | (____/\\| (____/\\___) (___| (____/\\| )  \\  |   | |   ")
-        println(" (_______/(_______/\\_______/(_______/|/    )_)   )_(   ")
-        println("_________________________________________________________")
-        // Please dont break this
-    }
-}
