--- src/main/java/dev/toastmc/client/util/EntityUtils.kt
@@ -1,54 +1,0 @@
-package dev.toastmc.client.util
-
-import net.minecraft.client.MinecraftClient
-import net.minecraft.entity.Entity
-import net.minecraft.entity.LivingEntity
-import net.minecraft.entity.Saddleable
-import net.minecraft.entity.mob.*
-import net.minecraft.entity.passive.AnimalEntity
-import net.minecraft.entity.passive.GolemEntity
-import net.minecraft.entity.passive.VillagerEntity
-import net.minecraft.entity.passive.WolfEntity
-import net.minecraft.entity.vehicle.BoatEntity
-import net.minecraft.entity.vehicle.MinecartEntity
-
-object EntityUtils {
-    private val mc = MinecraftClient.getInstance()
-    fun notSelf(e: Entity): Boolean {
-        return e !== mc.player && e !== mc.cameraEntity
-    }
-
-    fun isAnimal(e: Entity?): Boolean {
-        return e is AnimalEntity ||
-                e is AmbientEntity ||
-                e is WaterCreatureEntity ||
-                e is GolemEntity && !e.handSwinging ||
-                e is VillagerEntity
-    }
-
-    fun isLiving(e: Entity?): Boolean {
-        return e is LivingEntity
-    }
-
-    fun isHostile(e: Entity?): Boolean {
-        return e is HostileEntity && e !is PiglinEntity && e !is EndermanEntity ||
-                e is PiglinEntity && e.isAngryAt(mc.player) ||
-                e is WolfEntity && e.isAttacking && e.ownerUuid !== mc.player!!.uuid ||
-                e is EndermanEntity && e.isAngry ||
-                e is GolemEntity && e.handSwinging ||
-                e is MobEntity && e.isAttacking
-    }
-
-    fun isNeutral(e: Entity?): Boolean {
-        return e is PiglinEntity && !e.isAngryAt(mc.player) ||
-                e is WolfEntity && (!e.isAttacking || e.ownerUuid === mc.player!!.uuid) ||
-                e is EndermanEntity && !e.isAngry ||
-                e is GolemEntity && !e.handSwinging
-    }
-    fun isVehicle(e: Entity?): Boolean {
-        return e is BoatEntity ||
-                e is MinecartEntity ||
-                (e is Saddleable && e.isSaddled)
-    }
-
-}
